fix route delete w/ multer, end of the TP

+++ controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/thing');
+const fs = require ('fs');
 
 exports.createThing = (req, res, next) => {
   const thingObject = JSON.parse(req.body.thing);
@@ -42,9 +43,20 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({_id: req.params.id})
-  .then(() => res.status(200).json({message: 'Deleted!'}))
-  .catch(error => res.status(400).json({error}));
+  Thing.findOne({_id: req.params.id})
+  .then(thing => {
+    if (thing.userId != req.auth.userId) {
+      res.status(401).json({message: 'Non-autorisé'});
+    } else {
+      const filename = thing.imageUrl.split('/images/')[1];
+      fs.unlink(`image/${filename}`, () => {
+        Thing.deleteOne({_id: req.params.id})
+        .then(() => res.status(200).json({message: 'Objet supprimé !'}))
+        .catch(error => res.status(400).json({error}));
+      })
+    }
+  })
+  .catch(error => res.status(500).json({error}));
 };
 
 exports.getAllStuff = (req, res, next) => {
